Rotation, flipping and several image functions moved to xmipp3/utils for reuse
