+++ xmipp3/protocols/protocol_movie_correlation.py
@@ -27,10 +27,9 @@
 # *
 # **************************************************************************
 
-import os
-import math
-from math import ceil
+import os.path
 import numpy as np
+from math import ceil
 
 import pyworkflow.utils as pwutils
 from pyworkflow.utils import yellowStr
@@ -46,6 +45,7 @@ from pwem.protocols import ProtAlignMovies
 
 from xmipp3.convert import writeMovieMd
 from xmipp3.base import isXmippCudaPresent
+from ..utils import *
 
 
 class XmippProtMovieCorr(ProtAlignMovies):
@@ -230,43 +230,6 @@ class XmippProtMovieCorr(ProtAlignMovies):
                  2: np.asarray([[-1, 0, imag_array.shape[1]], [0, 1, 0], [0, 0, 1]])}
       return flipDic[self.gainFlip.get()]
 
-    def transformGain(self, gainFn, outFn=None):
-      '''Transforms the gain image with the user especifications'''
-      if outFn == None:
-        ext = pwutils.getExt(gainFn)
-        baseName = os.path.basename(gainFn).replace(ext, '_transformed' + ext)
-        outFn = os.path.abspath(self._getExtraPath(baseName))
-
-      gainImg = self.readImage(gainFn)
-      imag_array = np.asarray(gainImg.getData(), dtype=np.float64)
-
-      # Building the transformation matrix
-      angle = self.getUserAngle()
-      M = self.getUserFlip(imag_array)
-      print('Transforming gain: {} degrees rotation, {} flip'.
-            format(angle, ['no', 'vertical', 'horizontal'][self.gainFlip.get()]))
-      flipped_array, M = rotation(imag_array, angle, imag_array.shape, M)
-      self.writeImageFromArray(flipped_array, outFn)
-      return outFn
-
-    def flipYGain(self, gainFn, outFn=None):
-      '''Flips an image in the Y axis'''
-      if outFn == None:
-        ext = pwutils.getExt(gainFn)
-        if not '_flipped' in os.path.basename(gainFn):
-          baseName = os.path.basename(gainFn).replace(ext, '_flipped' + ext)
-        else:
-          baseName = os.path.basename(gainFn).replace('_flipped', '')
-        outFn = os.path.abspath(self._getExtraPath(baseName))
-      gainImg = self.readImage(gainFn)
-      imag_array = np.asarray(gainImg.getData(), dtype=np.float64)
-
-      #Flipped Y matrix
-      M, angle = np.asarray([[1, 0, 0], [0, -1, imag_array.shape[0]], [0, 0, 1]]), 0
-      flipped_array, M = rotation(imag_array, angle, imag_array.shape, M)
-      self.writeImageFromArray(flipped_array, outFn)
-      return outFn
-
     def getGPUArgs(self):
         args = ' --device %(GPU)s'
         if self.doLocalAlignment.get():
@@ -352,6 +315,43 @@ class XmippProtMovieCorr(ProtAlignMovies):
         self._saveAlignmentPlots(movie)
 
     #--------------------------- UTILS functions ------------------------------
+    def transformGain(self, gainFn, outFn=None):
+      '''Transforms the gain image with the user especifications'''
+      if outFn == None:
+        ext = pwutils.getExt(gainFn)
+        baseName = os.path.basename(gainFn).replace(ext, '_transformed' + ext)
+        outFn = os.path.abspath(self._getExtraPath(baseName))
+
+      gainImg = readImage(gainFn)
+      imag_array = np.asarray(gainImg.getData(), dtype=np.float64)
+
+      # Building the transformation matrix
+      angle = self.getUserAngle()
+      M = self.getUserFlip(imag_array)
+      print('Transforming gain: {} degrees rotation, {} flip'.
+            format(angle, ['no', 'vertical', 'horizontal'][self.gainFlip.get()]))
+      flipped_array, M = rotation(imag_array, angle, imag_array.shape, M)
+      writeImageFromArray(flipped_array, outFn)
+      return outFn
+
+    def flipYGain(self, gainFn, outFn=None):
+      '''Flips an image in the Y axis'''
+      if outFn == None:
+        ext = pwutils.getExt(gainFn)
+        if not '_flipped' in os.path.basename(gainFn):
+          baseName = os.path.basename(gainFn).replace(ext, '_flipped' + ext)
+        else:
+          baseName = os.path.basename(gainFn).replace('_flipped', '')
+        outFn = os.path.abspath(self._getExtraPath(baseName))
+      gainImg = readImage(gainFn)
+      imag_array = np.asarray(gainImg.getData(), dtype=np.float64)
+
+      #Flipped Y matrix
+      M, angle = np.asarray([[1, 0, 0], [0, -1, imag_array.shape[0]], [0, 0, 1]]), 0
+      flipped_array, M = rotation(imag_array, angle, imag_array.shape, M)
+      writeImageFromArray(flipped_array, outFn)
+      return outFn
+
     def _getShiftsFile(self, movie):
         return self._getExtraPath(self._getMovieRoot(movie) + '_shifts.xmd')
 
@@ -487,39 +487,3 @@ class XmippProtMovieCorr(ProtAlignMovies):
 
     def _citations(self):
         return ['strelak2020flexalign']
-
-
-    def writeImageFromArray(self, array, fn):
-        img = emlib.Image()
-        img.setData(array)
-        img.write(fn)
-
-    def readImage(self, fn):
-        img = emlib.Image()
-        img.read(fn)
-        return img
-
-
-def applyTransform(imag_array, M, shape):
-  ''' Apply a transformation(M) to a np array(imag) and return it in a given shape
-  '''
-  imag = emlib.Image()
-  imag.setData(imag_array)
-  imag = imag.applyWarpAffine(list(M.flatten()), shape, True)
-  return imag.getData()
-
-
-def rotation(imag, angle, shape, P):
-  '''Rotate a np.array and return also the transformation matrix
-  #imag: np.array
-  #angle: angle in degrees
-  #shape: output shape
-  #P: transform matrix (further transformation in addition to the rotation)'''
-  (hsrc, wsrc) = imag.shape
-  angle *= math.pi / 180
-  T = np.asarray([[1, 0, -wsrc / 2], [0, 1, -hsrc / 2], [0, 0, 1]])
-  R = np.asarray([[math.cos(angle), math.sin(angle), 0], [-math.sin(angle), math.cos(angle), 0], [0, 0, 1]])
-  M = np.matmul(np.matmul(np.linalg.inv(T), np.matmul(R, T)), P)
-
-  transformed = applyTransform(imag, M, shape)
-  return transformed, M
